// src/common/enums/env.type.ts
export type EnvListType = {
  PORT: string;

  NODE_ENV: string;

  HOST: string;

  POSTGRES_HOST: string;

  POSTGRES_PORT: string;

  POSTGRES_USER: string;

  POSTGRES_PASSWORD: string;

  POSTGRES_DATABASE: string;

  TOKEN_EXPIRATION: string;

  JWT_SECRET_KEY: string;

  PD_POSTGRES_HOST: string;

  PD_POSTGRES_PORT: string;

  PD_POSTGRES_USERNAME: string;

  PD_POSTGRES_PASSWORD: string;

  PD_POSTGRES_DATABASE: string;

  ENDPOINT_CORS: string;
};